requirements.txt
# Core dependencies
pymongo==4.4.1
python-multipart==0.0.5
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.34.0
pydantic==2.5.2
httpx>=0.24.1
prometheus-client==0.19.0
psutil==5.9.8
python-socketio>=5.8.0
aiohttp>=3.8.5

# File processing
python-docx==1.1.2
Pillow==11.1.0
pypdf==5.3.1
pymupdf==1.24.0
aiofiles==23.2.1
pdf2image==1.17.0

# Azure storage
azure-storage-blob==12.19.0

# AI/ML dependencies (minimal versions)
google-generativeai>=0.3.1

# Testing dependencies (optional for production)
# pytest>=7.0.0
# pytest-asyncio>=0.18.0
# pytest-cov>=2.12.0

# Development dependencies (optional for production)
# black>=22.0.0
# isort>=5.10.0
# mypy>=0.910